nusamai-geojson から nusamai-plateau への依存を切る
