+++ app/src-tauri/Cargo.toml
@@ -18,6 +18,7 @@ tauri-build = { version = "1.5.0", features = [] }
 serde_json = "1.0"
 serde = { version = "1.0", features = ["derive"] }
 tauri = { version = "1.5.2", features = ["dialog-all"] }
+nusamai = { path = "../../nusamai" }
 nusamai-geometry = { path = "../../nusamai-geometry" }
 nusamai-geojson = { path = "../../nusamai-geojson" }
 nusamai-plateau = { path = "../../nusamai-plateau" }
