+++ app/src-tauri/src/example.rs
@@ -2,7 +2,7 @@
 //! nusamai-geojson の exmaple/gml2geojson を元にした、暫定的な処理
 
 use citygml::{CityGMLElement, CityGMLReader, ParseError, SubTreeReader};
-use nusamai_geojson::toplevel_cityobj_to_geojson_features;
+use nusamai::sink::geojson::toplevel_cityobj_to_geojson_features;
 use nusamai_plateau::TopLevelCityObject;
 use std::fs;
 use std::io::BufRead;
