+++ nusamai/Cargo.toml
@@ -22,3 +22,4 @@ url = "2.5.0"
 
 [dev-dependencies]
 rand = "0.8.5"
+nusamai-geometry = { path = "../nusamai-geometry" }
