+++ nusamai/examples/gml2sliced_geojson.rs
@@ -1,4 +1,4 @@
-//! ポリゴンをタイル状にスライスするデモ
+//! ポリゴンをタイル状にスライスしてGeoJSONとして出力するデモ
 //!
 //! 使用例:
 //!
